Move exceptions to exception folder
