+++ src/main/java/vn/unigap/api/exception/ApiException.java
@@ -1,0 +1,15 @@
+package vn.unigap.api.exception;
+
+public class ApiException {
+    public static class BadInputException extends RuntimeException {
+        public BadInputException(String message) {
+            super(message);
+        }
+    }
+
+    public static class NotFoundException extends RuntimeException {
+        public NotFoundException(String message) {
+            super(message);
+        }
+    }
+}
